requirements.txt
absl-py==2.1.0
asgiref==3.8.1
astunparse==1.6.3
certifi==2024.12.14
charset-normalizer==3.4.1
Django==5.1.5
django-cors-headers==4.6.0
django-environ==0.12.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
flatbuffers==25.1.24
gast==0.6.0
google-pasta==0.2.0
grpcio==1.70.0
gunicorn==23.0.0
h5py==3.12.1
idna==3.10
keras==3.8.0
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml-dtypes==0.3.2
namex==0.0.8
numpy==1.26.4
opt_einsum==3.4.0
optree==0.14.0
packaging==24.2
pillow==11.1.0
protobuf==4.25.6
psycopg2==2.9.10
psycopg2-binary==2.9.10
Pygments==2.19.1
PyJWT==2.10.1
python-dotenv==1.0.1
requests==2.32.3
rich==13.9.4
six==1.17.0
sqlparse==0.5.3
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.2
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
whitenoise==6.8.2
wrapt==1.17.2
